+++ src/funcs/queryExportSemanticThreadsQueryResults.ts
@@ -30,7 +30,7 @@ import { Result } from "../types/fp.js";
  */
 export function queryExportSemanticThreadsQueryResults(
   client: InkeepAnalyticsCore,
-  request: components.QuerySemanticThreadsParamsSchema,
+  request: components.QuerySemanticThreadsParams,
   options?: RequestOptions,
 ): APIPromise<
   Result<
@@ -58,7 +58,7 @@ export function queryExportSemanticThreadsQueryResults(
 
 async function $do(
   client: InkeepAnalyticsCore,
-  request: components.QuerySemanticThreadsParamsSchema,
+  request: components.QuerySemanticThreadsParams,
   options?: RequestOptions,
 ): Promise<
   [
@@ -83,7 +83,7 @@ async function $do(
   const parsed = safeParse(
     request,
     (value) =>
-      components.QuerySemanticThreadsParamsSchema$outboundSchema.parse(value),
+      components.QuerySemanticThreadsParams$outboundSchema.parse(value),
     "Input validation failed",
   );
   if (!parsed.ok) {